+++ 1.11.2/src/main/java/com/stateshifterlabs/achievementbooks/commands/ReloadCommand.java
@@ -5,6 +5,7 @@ import net.minecraft.command.CommandBase;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.util.text.TextComponentString;
 
 public class ReloadCommand extends CommandBase {
 
@@ -28,6 +29,8 @@ public class ReloadCommand extends CommandBase {
 	@Override
 	public void execute(MinecraftServer server, ICommandSender sender, String[] args) throws CommandException {
 		loader.init();
+		String message = "Achievement Books Data reloaded. Textures may not appear correctly until a full game restart";
+		sender.sendMessage(new TextComponentString(message));
 	}
 
 }