+++ 1.7.10/src/main/java/com/stateshifterlabs/achievementbooks/commands/ReloadCommand.java
@@ -3,6 +3,7 @@ package com.stateshifterlabs.achievementbooks.commands;
 import com.stateshifterlabs.achievementbooks.data.Loader;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.ICommandSender;
+import net.minecraft.util.ChatComponentText;
 
 public class ReloadCommand extends CommandBase {
 
@@ -24,7 +25,9 @@ public class ReloadCommand extends CommandBase {
 	}
 
 	@Override
-	public void processCommand(ICommandSender p_71515_1_, String[] p_71515_2_) {
+	public void processCommand(ICommandSender sender, String[] p_71515_2_) {
 		loader.init();
+		String message = "Achievement Books Data reloaded. Textures may not appear correctly until a full game restart";
+		sender.addChatMessage(new ChatComponentText(message));
 	}
 }